py-simputils-config-9 Enums for keys and argparser.Namespace proper support
- Improved type hints and typing in general

closes #9
