+++ src/simputils/config/components/handlers/DotEnvFileHandler.py
@@ -1,12 +1,12 @@
 import os
 from io import IOBase
-from os import PathLike
 
 import dotenv
 
 from simputils.config.enums import ConfigStoreType
 from simputils.config.generic import BasicFileHandler
 from simputils.config.models import ConfigStore
+from simputils.config.types import FileType
 
 
 class DotEnvFileHandler(BasicFileHandler):
@@ -16,12 +16,10 @@ class DotEnvFileHandler(BasicFileHandler):
 
 	CONFIG_TYPE: str = ConfigStoreType.DOT_ENV
 
-	def process_file(self, file: PathLike | str | IOBase) -> ConfigStore | None:
+	def process_file(self, file: FileType) -> ConfigStore | None:
 		conf = self._prepare_conf(file)
 		if conf is not None:
 
-			# noinspection PyBroadException
-			# try:
 			if isinstance(file, IOBase):
 				data = dotenv.dotenv_values(stream=file)
 				return conf.config_apply(
@@ -40,7 +38,5 @@ class DotEnvFileHandler(BasicFileHandler):
 					type=conf.type,
 					handler=self,
 				)
-			# except Exception:  # pragma: no cover
-			# 	pass
 
 		return None  # pragma: no cover
