+++ src/simputils/config/generic/BasicAppliedConf.py
@@ -1,0 +1,4 @@
+
+
+class BasicAppliedConf:
+	pass
