+++ src/simputils/config/models/AppliedConf.py
@@ -1,16 +1,19 @@
 from dataclasses import dataclass
 from typing import Any
 
+from simputils.config.generic import BasicAppliedConf
+from simputils.config.types import SourceType, HandlerType
+
 
 @dataclass
-class AppliedConf:
+class AppliedConf(BasicAppliedConf):
 	"""
 	ConfigStore's history record
 	"""
 
-	applied_keys: list = None
+	applied_keys: list[str] = None
 	type: str = None
 	name: str = None
-	source: str = None
-	handler: Any = None
-	ref: dict = None
+	source: SourceType = None
+	handler: HandlerType = None
+	ref: Any = None
