+++ src/simputils/config/models/__init__.py
@@ -1,2 +1,3 @@
+from .AnnotatedConfigData import AnnotatedConfigData
 from .AppliedConf import AppliedConf
 from .ConfigStore import ConfigStore
